Update RO_FASA_Saturn.cfg
I know this was only recently changed.  I'm also aware that the Saturn V traditionally only fired the J-2 on the SIV-B stage twice.  But according to Wikipedia and other references, the only limitation on how many times the J-2 could be restarted was fuel supply and the amount of Helium available for tank pressurization.
The main difference between the S-IVB-200 and S-IVB-500 (besides the flared fairing) was that the -200 series did not include as much Helium since it would not need to be restarted in flight.
According to http://www.nasa.gov/centers/marshall/pdf/499245main_J2_Engine_fs.pdf, page 6-7, "Refill of the gaseous helium tank was not required because the original ground-fill supply was sufficient for three starts".

Furthermore, the reference document states (http://www.nasa.gov/centers/marshall/pdf/499245main_J2_Engine_fs.pdf page 6-2) "The ASI (Augmented Spark Igniter) operates continuously during entire engine firing, is uncooled, and is capable of multiple reignitions under all environmental conditions."

Plus, the only in-game way to simulate the fuel expelled that allowed the S-IVB to be redirected for lunar impact is with an additional burn.